+++ app/controllers/patients/pipelines/computations_controller.rb
@@ -59,7 +59,11 @@ module Patients
 
       def prepare_to_show_computation
         @computations = @pipeline.computations.order(:created_at)
-        @details = Patients::Details.new(@patient.case_number, current_user).call
+
+        unless request.xhr?
+          @details = Patients::Details.
+                     new(@patient.case_number, current_user).call
+        end
 
         if load_versions?
           @versions = Gitlab::Versions.