+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 - Default segmentation run mode changed into 3 (@mkasztelnik)
 - Patient.case_number now used as patient ID in HTTP requests (@amber7b)
 - Computation update interval changed into 30 seconds (@mkasztelnik)
+- Load patient details only when needed (@mkasztelnik)
 
 ### Deprecated